requirements.txt
torch
torchvision
numpy
pandas
Pillow
tqdm
hydra-core
omegaconf
scikit-learn
datasets
easydict
PyYAML
requests
gdown
opencv-python
webcolors
open_clip_torch
transformers
huggingface_hub
git+https://github.com/openai/CLIP.git
